图片截图蒙层、添加step2
