+++ packages/theme/pages/step2.vue
@@ -1,0 +1,459 @@
+<template>
+  <div id="form-step2">
+    <SfNotification
+      :visible="Boolean(notificationVisible)"
+      persistent=""
+      title=""
+      :message="notificationVisible"
+      action=""
+      type="danger"
+    />
+    <h3 class="form__h2">Wig Specs</h3>
+    <form class="form">
+      <SfComponentSelect
+        v-model="density"
+        label="Hair Density "
+        class="form__element form__element--half form__element--half-even form__select sf-component-select--underlined"
+        :required="false"
+        valid
+        :disabled="false"
+        error-message="Please select Hair Density "
+        placeholder="Please select Hair Density "
+      >
+        <SfComponentSelectOption
+          v-for="density in densities"
+          :key="density.value"
+          :value="density.value"
+        >
+          {{ density.label }}
+        </SfComponentSelectOption>
+      </SfComponentSelect>
+      <SfComponentSelect
+        v-model="laceMaterial"
+        label="Lace Material"
+        class="form__element form__element--half form__select sf-component-select--underlined"
+        :required="false"
+        valid
+        :disabled="false"
+        error-message="Please select Lace material."
+        placeholder="Please select Lace material."
+      >
+        <SfComponentSelectOption
+          v-for="item in laceMaterials"
+          :key="item.value"
+          :value="item.value"
+        >
+          {{ item.label }}
+        </SfComponentSelectOption>
+      </SfComponentSelect>
+
+      <SfComponentSelect
+        v-model="cap"
+        label="Cap Construction "
+        class="form__element form__element--half form__element--half-even form__select sf-component-select--underlined"
+        :required="false"
+        valid
+        :disabled="false"
+        error-message="Please select Cap Construction "
+        placeholder="Please select Cap Construction "
+      >
+        <SfComponentSelectOption
+          v-for="item in caps"
+          :key="item.value"
+          :value="item.value"
+        >
+          {{ item.label }}
+        </SfComponentSelectOption>
+      </SfComponentSelect>
+
+      <SfComponentSelect
+        v-model="hairLine"
+        label="Hair Line"
+        class="form__element form__element--half form__select sf-component-select--underlined"
+        :required="false"
+        valid
+        :disabled="false"
+        error-message="Please select HairLine"
+        placeholder="Please select HairLine"
+      >
+        <SfComponentSelectOption
+          v-for="(item, key) in hairLines"
+          :key="key"
+          :value="item.value"
+        >
+          {{ item.label }}
+        </SfComponentSelectOption>
+      </SfComponentSelect>
+
+      <SfComponentSelect
+        v-model="capSize"
+        label="Cap Size "
+        class="form__element form__element--half form__element--half-even form__select sf-component-select--underlined"
+        :required="false"
+        valid
+        :disabled="false"
+        error-message="Please select Cap Size "
+        placeholder="Please select Cap Size "
+      >
+        <SfComponentSelectOption
+          v-for="item in capSizes"
+          :key="item.value"
+          :value="item.value"
+        >
+          {{ item.label }}
+        </SfComponentSelectOption>
+      </SfComponentSelect>
+      <SfComponentSelect
+        v-model="addElasticBand"
+        label="Add Elastic Bands"
+        class="form__element form__element--half form__select sf-component-select--underlined"
+        :required="false"
+        valid
+        :disabled="false"
+        error-message="Please select Add Elastic Bands"
+        placeholder="Please select Add Elastic Bands"
+      >
+        <SfComponentSelectOption
+          v-for="(item, key) in addElasticBands"
+          :key="key"
+          :value="item.value"
+        >
+          {{ item.label }}
+        </SfComponentSelectOption>
+      </SfComponentSelect>
+
+      <div class="form__element form__element--half" />
+
+      <div class="form__action">
+        <SfButton type="submit" @click.prevent="submit">Next</SfButton>
+      </div>
+    </form>
+    <div class="pdc-pdp" v-if="isLoadervisible">
+      <SfLoader class="pdc-pdp-loader" :loading="isLoadervisible">
+        <div class="desc">
+          please have a cup of coffee,it will be done in one or wait minutes
+        </div>
+      </SfLoader>
+      <div class="pdc-pdp-desc">
+        please have a cup of coffee,it will be done in one or wait minutes
+      </div>
+    </div>
+  </div>
+</template>
+<script>
+import('@google/model-viewer');
+
+import {
+  SfSelect,
+  SfColor,
+  SfButton,
+  SfInput,
+  SfComponentSelect,
+  SfHeading,
+  SfLoader,
+  SfNotification,
+  SfIcon,
+  SfSidebar,
+  SfImage,
+} from '@storefront-ui/vue';
+
+export default {
+  name: 'Step1',
+  components: {
+    SfSelect,
+    SfColor,
+    SfButton,
+    SfInput,
+    SfComponentSelect,
+    SfHeading,
+    SfLoader,
+    SfNotification,
+    SfIcon,
+    SfSidebar,
+    SfImage,
+  },
+  // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+  setup(props, { root }) {
+    const handleNextClick = () => {
+      return root.$router.push({
+        path: '/model',
+        query: {
+          path: this.filePath,
+        },
+      });
+    };
+    return {
+      handleNextClick,
+    };
+  },
+  // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+  data() {
+    return {
+      isLoadervisible: false,
+      valid: false,
+      submitted: false,
+      notificationVisible: '',
+      density: '150%',
+      laceMaterial: 'normalLace',
+      cap: '4',
+      hairLine: 'naturalHairLine',
+      capSize: 'average',
+      addElasticBand: 'no',
+      densities: [
+        { label: '150%', value: '150%' },
+        { label: '180% +$30.00', value: '180%' },
+      ],
+      laceMaterials: [
+        { label: 'HD Lace +$20.00', value: 'hdLace' },
+        { label: 'Normal Lace', value: 'normalLace' },
+      ],
+      caps: [
+        { label: '4 Parting Glueless Lace Front Crap', value: '4' },
+        {
+          label: '6 Deep Parting Glueless Lace Front Crap +$60.00 GlueLess',
+          value: '6',
+        },
+        { label: '5 * 5 Closure Lace Cap +$40.00', value: '5' },
+      ],
+      hairLines: [
+        { label: 'Natural Hair Line', value: 'naturalHairLine' },
+        { label: 'Pre-plucked HairLine', value: 'prePluckedHairLine' },
+      ],
+      capSizes: [
+        { label: 'Average', value: 'average' },
+        { label: 'Petite', value: 'petite' },
+        { label: 'Large', value: 'large' },
+        { label: 'Custom +$30.00', value: 'custom' },
+      ],
+      addElasticBands: [
+        { label: 'Yes', value: 'yes' },
+        { label: 'No', value: 'no' },
+      ],
+      // 轮询时间
+      timer: null,
+      // 是否执行轮训
+      is2D: '',
+      requestId: '',
+      // 回显图片路径
+      filePath: '',
+    };
+  },
+  watch: {
+    requestId: {
+      // 查看文件上传的处理状态
+      handler(newVal) {
+        if (newVal !== '' && this.is2D !== 'done') {
+          // 实现轮询
+          this.createSetInterval();
+        } else if (newVal !== '' && this.is2D === 'done') {
+          this.stopSetInterval();
+        }
+      },
+      immediate: true,
+    },
+  },
+  methods: {
+    async submit() {
+      const info = JSON.parse(window.localStorage.getItem('info'));
+      const hairInfo = JSON.parse(window.localStorage.getItem('hairInfo'));
+      const params = [
+        ...hairInfo,
+        this.density,
+        this.laceMaterial,
+        this.cap,
+        this.hairLine,
+        this.capSize,
+        this.addElasticBand,
+      ];
+      const newData = {
+        name: info.name,
+        params,
+        data: info.data,
+      };
+
+      // this.$store.dispatch('addForm', data);
+      this.submitted = true;
+      this.isLoadervisible = true;
+      // 获取远端图片
+      this.notificationVisible = '';
+      await this.$axios({
+        method: 'POST',
+        // url: '/ama/profile',
+        url: '/b/default/profile',
+        data: JSON.stringify(newData),
+      })
+        .then(({ data }) => {
+          window.localStorage.removeItem('filePath');
+          this.requestId = data.request_id;
+          this.filePath = data.file_path;
+        })
+        .catch((e) => {
+          console.log(e);
+          this.notificationVisible = 'Internal Server Error';
+          this.isLoadervisible = false;
+        });
+    },
+    // reset() {
+    //   this.style = '';
+    //   this.length = '';
+    //   this.color = '';
+    //   this.density = '';
+    //   this.laceMaterial = '';
+    // }
+    // 开启轮询  如果存在则先销毁定时器后重新开启
+    // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+    createSetInterval() {
+      this.stopSetInterval();
+      this.timer = setInterval(() => {
+        this.getNewMessage();
+      }, 5000);
+    },
+    // 关闭轮询
+    // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+    stopSetInterval() {
+      if (this.timer) {
+        clearInterval(this.timer);
+        this.timer = null;
+      }
+    },
+    // 请求是否有新消息
+    // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+    async getNewMessage() {
+      await this.$axios({
+        method: 'GET',
+        // url: '/ama/status',
+        url: '/b/default/status',
+        headers: {
+          'x-jizhan-request-id': this.requestId,
+        },
+      })
+        .then(({ data }) => {
+          if (data.status === 'done') {
+            this.isLoadervisible = false;
+            this.stopSetInterval();
+            this.$router.push({
+              path: '/model',
+              // query: {
+              //   path: this.filePath
+              // }
+            });
+            window.localStorage.setItem('filePath', this.filePath);
+          } else if (data.status === 'timeout') {
+            this.notificationVisible = '处理超时，请重试';
+            this.isLoadervisible = false; // 选择配置的暂时不支持，请重新配置
+            this.stopSetInterval();
+          } else if (data.status === 'bad') {
+            this.notificationVisible = ' 选择配置的暂时不支持，请重新配置';
+            this.isLoadervisible = false;
+            this.stopSetInterval();
+          }
+          this.is2D = data.status;
+        })
+        .catch((e) => {
+          console.log(e);
+          this.notificationVisible = 'Internal Server Error';
+          this.isLoadervisible = false;
+          this.stopSetInterval();
+        });
+    },
+  },
+  mounted() {
+    this.stopSetInterval();
+  },
+};
+</script>
+
+<style lang="scss" scoped>
+@import '~@storefront-ui/vue/styles';
+
+.pdc-pdp {
+  min-height: 93vh;
+  width: 100%;
+  position: absolute;
+  top: 0;
+  right: 0;
+  z-index: 99999;
+  margin-left: -25%;
+  background: rgba(94, 91, 91, 0.2);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  @include for-mobile {
+    min-height: 165vh;
+  }
+  .pdc-pdp-loader {
+    width: 100%;
+    .sf-loader__overlay {
+      background: rgba(0, 0, 0, 0.5);
+    }
+  }
+  .pdc-pdp-desc {
+    color: red;
+    margin: var(--spacer-2xl);
+    @include for-mobile {
+      margin: var(--spacer-2xl) var(--spacer-base);
+    }
+  }
+}
+#form-ste2 {
+  box-sizing: border-box;
+  padding: 0 var(--spacer-sm);
+  position: relative;
+  @include for-desktop {
+    padding: 0 var(--spacer-sm);
+    max-width: 870px;
+    margin: 0 auto;
+  }
+}
+.form__h2 {
+  padding: var(--spacer-sm);
+}
+.form {
+  padding: var(--spacer-sm) 0;
+  @include for-mobile {
+    padding: var(--spacer-sm);
+  }
+
+  &__group {
+    display: flex;
+    align-items: flex-start;
+  }
+  &__action-button {
+    &:first-child {
+      margin: var(--spacer-sm) 0 0 0;
+    }
+    &--secondary {
+      margin: var(--spacer-sm) 0 var(--spacer-sm) var(--spacer-base);
+    }
+  }
+  &__button {
+    --button-width: 100%;
+  }
+  @include for-desktop {
+    display: flex;
+    flex-wrap: wrap;
+    align-items: flex-start;
+    margin: 0 var(--spacer-2xl) 0 0;
+    &:last-of-type {
+      margin: 0 calc(var(--spacer-2xl) - var(--spacer-sm)) 0 0;
+    }
+    &__action {
+      flex: 0 0 100%;
+      display: flex;
+    }
+    &__button {
+      --button-width: auto;
+    }
+  }
+  .sf-loader {
+    position: absolute;
+    top: 0;
+    left: 0;
+    z-index: 1;
+    .sf-loader__overlay {
+      background: rgba(0, 0, 0, 0.7);
+    }
+  }
+}
+</style>
